test: generalize exec-server fixture
