+++ codex-rs/exec-server/testing/exec_server.rs
@@ -1,0 +1,14 @@
+//! Minimal exec-server fixture for Bazel-only integration tests.
+//!
+//! Linking only exec-server avoids depending on the full Codex CLI binary
+//! when a test only needs a WebSocket executor endpoint.
+
+use codex_exec_server::ExecServerRuntimePaths;
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    let current_exe = std::env::current_exe()?;
+    let runtime_paths =
+        ExecServerRuntimePaths::new(current_exe, /*codex_linux_sandbox_exe*/ None)?;
+    codex_exec_server::run_main("ws://127.0.0.1:0", runtime_paths).await
+}
